supprime tout de suite les identifiants d'API des opérations dont le récupérateur a été supprimé

+++ app/services/destroy_bank_accounts.rb
@@ -5,10 +5,7 @@ class DestroyBankAccounts
 
   def execute
     @bank_accounts.each do |bank_account|
-      unless bank_account.retriever
-        bank_account.operations.update_all(api_id: nil)
-        bank_account.destroy
-      end
+      bank_account.destroy unless bank_account.retriever
     end
   end
 end
